Update minimum supported go version to 1.11 Add support for go 1.13
seprate out the flags to individual files

fix md2man dependency issue
