+++ .travis.yml
@@ -3,9 +3,9 @@ sudo: false
 dist: trusty
 osx_image: xcode8.3
 go:
-  - 1.10.x
   - 1.11.x
   - 1.12.x
+  - 1.13.x
 
 os:
   - linux
@@ -16,16 +16,11 @@ cache:
     - node_modules
 
 before_script:
-  - go get github.com/shurcooL/vfsgen || true
-  - go get github.com/shurcooL/httpfs/union || true
   - go get github.com/urfave/gfmrun/... || true
   - go get golang.org/x/tools/cmd/goimports
-  - if [ ! -f node_modules/.bin/markdown-toc ] ; then
-      npm install markdown-toc ;
-    fi
+  - npm install markdown-toc ;
 
 script:
-  - go run build.go generate
   - go run build.go vet
   - go run build.go test
   - go run build.go gfmrun
