+++ go.mod
@@ -1,9 +1,9 @@
 module github.com/urfave/cli
 
-go 1.10
+go 1.11
 
 require (
 	github.com/BurntSushi/toml v0.3.1
-	github.com/cpuguy83/go-md2man v1.0.10
+	github.com/cpuguy83/go-md2man/v2 v2.0.0-20190314233015-f79a8a8ca69d
 	gopkg.in/yaml.v2 v2.2.2
 )
